[chg]Device::vibrate on ios.

+++ cocos/platform/ios/CCDevice-ios.mm
@@ -608,7 +608,7 @@ void Device::setKeepScreenOn(bool value)
  @brief Only works on iOS devices that support vibration (such as iPhone). Should only be used for important alerts. Use risks rejection in iTunes Store.
  @param duration ignored for iOS
  */
-void Device::vibrate(float duration)
+void Device::vibrate(uint32_t duration)
 {
     // See http://stackoverflow.com/questions/4724980/making-the-iphone-vibrate
     // should vibrate no matter it is silient or not
